simulators/requirements.txt
paho-mqtt>=1.6.1
pika>=1.3.2
python-dotenv>=1.0.0
Faker>=20.0.0
requests>=2.31.0
